requirements.txt
CyberGearDriver==0.1.1
pyqtgraph==0.13.7
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
python-can==4.5.0
gs-usb==0.3.0
typing_extensions==4.12.2
